Rename module natfrp -> sakurafrp
