+++ tests/test_sakurafrp.py
@@ -2,7 +2,7 @@ import os
 import unittest
 from unittest.mock import patch
 
-from natfrp.natfrp import (
+from sakurafrp.sakurafrp import (
     NatfrpConfig,
     NatfrpError,
     _notify_if_configured,
@@ -22,7 +22,7 @@ class ConfigTest(unittest.TestCase):
                 NatfrpConfig.from_env()
 
 class NotificationTest(unittest.TestCase):
-    @patch("natfrp.natfrp.send_source_notification")
+    @patch("sakurafrp.sakurafrp.send_source_notification")
     def test_notifications_are_enabled_by_default(self, send_notification):
         values = {"TELEGRAM_TOKEN": "token", "TELEGRAM_CHAT_ID": "chat"}
         with patch.dict(os.environ, values, clear=True):
@@ -30,7 +30,7 @@ class NotificationTest(unittest.TestCase):
 
         send_notification.assert_called_once_with("SAKURAFRP", "test")
 
-    @patch("natfrp.natfrp.send_source_notification")
+    @patch("sakurafrp.sakurafrp.send_source_notification")
     def test_notifications_can_be_disabled(self, send_notification):
         values = {
             "SAKURAFRP_NOTIFY": "false",
